Corrected the size of the supported MSEX versions count in client information packet.

+++ Acn/Citp/Packets/Msex/CitpMsexClientInformation.cs
@@ -55,7 +55,7 @@ namespace Citp.Packets.Msex
         {
             base.WriteData(data);
 
-            data.Write(SupportedMSEXVersions.Count);
+            data.Write((byte)SupportedMSEXVersions.Count);
 
             foreach (ushort version in SupportedMSEXVersions)
                 data.Write(version);
